requirements.txt
amqp==5.1.1
beautifulsoup4==4.9.3
celery==5.2.7
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
requests==2.27.1
selenium==4.6.0
urllib3==1.26.5
python-dotenv==0.21.0
librabbitmq==2.0.0
